Seperated logic for each bug into its own file.
