+++ gameLogic/bugLogic/ant.logic.js
@@ -1,0 +1,7 @@
+export function checkAntMove(move) {
+    let bIsMoveGood = true
+    return bIsMoveGood
+    // a little different, need to check that end state doesn't break continuity
+}
+
+export default checkAntMove;
