+++ gameLogic/bugLogic/beetle.logic.js
@@ -1,0 +1,7 @@
+export function checkBeetleMove(move) {
+    let bIsMoveGood = true
+    return bIsMoveGood
+    //similar to queen bee, just need to include the on top case
+}
+
+export default checkBeetleMove;
