+++ gameLogic/bugLogic/spider.logic.js
@@ -1,0 +1,9 @@
+
+function checkSpiderMove(move) {
+    // also need to check that end state doesn't break continuity, also need to check that end state is three "hexes" away
+    // no necessarily 3 sides away
+    let bIsMoveGood = true
+    return bIsMoveGood
+}
+
+export default checkSpiderMove;
