+++ gameLogic/checkMove.js
@@ -1,4 +1,9 @@
 import Bug from "../gameObjects/bugs.js"
+import checkQueenBeeMove from "./bugLogic/queenBee.logic.js";
+import checkBeetleMove from "./bugLogic/beetle.logic.js";
+import checkHopperMove from "./bugLogic/hopper.logic.js";
+import checkAntMove from "./bugLogic/ant.logic.js";
+import checkSpiderMove from "./bugLogic/spider.logic.js";
 
 function isHexOccupied(move) { //is the space where the player intends to move occupied already
     let bIsHexOccupied = false
@@ -44,89 +49,3 @@ export function checkMove(move) {  //main function for checking movement logic, 
 }
 
 export default checkMove;
-
-function areBugsAdjacent(move) {
-    return(move.moveBug.isAdjacentToBug(move.destBug))
-}
-
-function getSingleHexMoveLength(moveBug, staticBug, destIndex) {
-    //need protection here if staticBug doesn't have an entry matching movebug?
-    return(Math.abs(destIndex - staticBug.adjacentArray.indexOf(moveBug)))
-}
-
-function isMoveOneHex(move) {
-    const moveLength = getSingleHexMoveLength(move.moveBug, move.destBug, move.destIndex)
-    if( moveLength == 1 || moveLength == 5) return true //equal to 5 only when moving from 0 to 5, or 5 to 0
-    else return false
-}
-
-function isMoveAStraightHop(move) {
-    //trace bugs in the line using their adjacent arrays
-    var isMoveGood = false
-    var currentBug = move.moveBug
-    var nextBug = move.moveBug.adjacentArray[move.destIndex]
-    const moveLength = getSingleHexMoveLength(currentBug, nextBug, move.destIndex)
-
-    //can find a cleaner way of doing this
-    if(moveLength == 3) {
-        if(nextBug == move.destBug) {
-            isMoveGood = true
-        }
-        else {
-            isMoveGood = true
-            while (currentBug != move.destBug) {
-               
-                if(nextBug === null) {
-                    isMoveGood = false
-                    break
-                }
-                currentBug = nextBug
-                if(currentBug == move.destBug)
-                    break
-                nextBug = currentBug.adjacentArray[move.destIndex]
-                
-            }
-        }
-    }
-    return isMoveGood
-}
-
-//could maybe simplfy/consolidate to "checkSingleHexMove" and use that for the beetle as well
-function checkQueenBeeMove(move) {
-    let bIsMoveGood = false
-    if(areBugsAdjacent(move)) {
-        bIsMoveGood = isMoveOneHex(move)
-    //} else {
-      //  bIsMoveGood = false
-    }
-    return bIsMoveGood
-    //check if move is to another side on the same bug
-    //if not then move must be on a side of the adjacent bug, the queen will also be on the adjacence list of that bug, so all we have to do 
-    //check to see if that new side is "one away"
-}
-
-function checkBeetleMove(move) {
-    let bIsMoveGood = true
-    return bIsMoveGood
-    //similar to queen bee, just need to include the on top case
-}
-
-function checkHopperMove(move) {
-    let bIsMoveGood = false
-    bIsMoveGood = isMoveAStraightHop(move)
-    return bIsMoveGood
-    //check that all sides in a straight line through the move have bugs attached
-}
-
-function checkSpiderMove(move) {
-    // also need to check that end state doesn't break continuity, also need to check that end state is three "hexes" away
-    // no necessarily 3 sides away
-    let bIsMoveGood = true
-    return bIsMoveGood
-}
-
-function checkAntMove(move) {
-    let bIsMoveGood = true
-    return bIsMoveGood
-    // a little different, need to check that end state doesn't break continuity
-}
